feat:위스퍼 api 파일 구조 변경
