--- src/main/java/org/mansumugang/mansumugang_service/openaiclient/OpenAIClient.java
@@ -1,24 +1,0 @@
-package org.mansumugang.mansumugang_service.openaiclient;
-
-import org.mansumugang.mansumugang_service.model.request.ChatGPTRequest;
-import org.mansumugang.mansumugang_service.model.request.WhisperTranscriptionRequest;
-import org.mansumugang.mansumugang_service.model.response.ChatGPTResponse;
-import org.mansumugang.mansumugang_service.model.response.WhisperTranscriptionResponse;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-
-@FeignClient(
-        name = "openai-service",
-        url = "${openai-service.urls.base-url}",
-        configuration = OpenAIClientConfig.class
-)
-public interface OpenAIClient {
-
-    @PostMapping(value = "${openai-service.urls.chat-url}", headers = {"Content-Type=application/json"})
-    ChatGPTResponse chat(@RequestBody ChatGPTRequest chatGPTRequest);
-
-    @PostMapping(value = "${openai-service.urls.create-transcription-url}", headers = {"Content-Type=multipart/form-data"})
-    WhisperTranscriptionResponse createTranscription(@ModelAttribute WhisperTranscriptionRequest whisperTranscriptionRequest);
-}
